Refactoring custom evaluators
