+++ tools/accuracy_checker/openvino/tools/accuracy_checker/evaluators/custom_evaluators/asr_encoder_prediction_joint_evaluator.py
@@ -21,7 +21,6 @@ import numpy as np
 
 from ...adapters import create_adapter
 from ...config import ConfigError
-from ...launcher import create_launcher
 from ...utils import contains_all, contains_any, read_pickle, get_path
 from ...logging import print_info
 from .asr_encoder_decoder_evaluator import AutomaticSpeechRecognitionEvaluator
@@ -30,12 +29,7 @@ from .asr_encoder_decoder_evaluator import AutomaticSpeechRecognitionEvaluator
 class ASREvaluator(AutomaticSpeechRecognitionEvaluator):
     @classmethod
     def from_configs(cls, config, delayed_model_loading=False, orig_config=None):
-        dataset_config = config['datasets']
-        launcher_config = config['launchers'][0]
-        if launcher_config['framework'] == 'dlsdk' and 'device' not in launcher_config:
-            launcher_config['device'] = 'CPU'
-
-        launcher = create_launcher(launcher_config, delayed_model_loading=True)
+        dataset_config, launcher, _ = cls.get_dataset_and_launcher_info(config)
         model = ASRModel(
             config.get('network_info', {}), launcher, config.get('_models', []), config.get('_model_is_blob'),
             delayed_model_loading
